+++ app/controllers/crudify/api/v1/api_keys_controller.rb
@@ -1,4 +1,4 @@
-module DeviseCrud
+module CRUDify
   module Api
     module V1
       class ApiKeysController < ApplicationController