+++ app/controllers/crudify/test_controller.rb
@@ -1,9 +1,9 @@
-module DeviseCrud
+module CRUDify
   module Api
     module V1
       class TestController < ApplicationController
         def index
-          render json: { message: "Hello from DeviseCrud Engine" }
+          render json: { message: "Hello from CRUDify Engine" }
         end
       end
     end