+++ app/jobs/crudify/application_job.rb
@@ -1,4 +1,4 @@
-module DeviseCrud
+module CRUDify
   class ApplicationJob < ActiveJob::Base
   end
 end