+++ app/mailers/crudify/application_mailer.rb
@@ -1,4 +1,4 @@
-module DeviseCrud
+module CRUDify
   class ApplicationMailer < ActionMailer::Base
     default from: "from@example.com"
     layout "mailer"