+++ app/models/crudify/application_record.rb
@@ -1,4 +1,4 @@
-module DeviseCrud
+module CRUDify
   class ApplicationRecord < ActiveRecord::Base
     self.abstract_class = true
   end