+++ lib/generators/crudify/migration/templates/create_api_keys.rb
@@ -1,6 +1,6 @@
 class CreateApiKeys < ActiveRecord::Migration[6.1]
   def change
-    create_table :devise_crud_api_keys do |t|
+    create_table :crudify_api_keys do |t|
       t.string :key, null: false, unique: true
       t.datetime :expires_at, null: false
       t.integer :status, default: 0, null: false
@@ -8,7 +8,7 @@ class CreateApiKeys < ActiveRecord::Migration[6.1]
       t.timestamps
     end
 
-    add_index :devise_crud_api_keys, :key, unique: true
-    add_index :devise_crud_api_keys, :status
+    add_index :crudify_api_keys, :key, unique: true
+    add_index :crudify_api_keys, :status
   end
 end