+++ CHANGELOG.md
@@ -2,4 +2,4 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
-- Initial release of `devise_crud` gem - yet to be planned
+- Initial release of `crudify` gem - yet to be planned